requirements.txt
Flask==3.1.0
python-docx==1.1.2
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
